requirements.txt
about-time==4.2.1
alive-progress==3.1.5
grapheme==0.6.0
pillow==10.2.0
Pygments==2.17.2
